UIU-1973 refactor: move user edit form sections to form dir
