+++ src/views/UserEdit/sections/EditContactInfo/EditContactInfo.js
@@ -15,8 +15,8 @@ import {
 } from '@folio/stripes/components';
 import { AddressEditList } from '@folio/stripes/smart-components';
 
-import { toAddressTypeOptions } from '../../data/converters/address_type';
-import contactTypes from '../../data/static/contactTypes';
+import { toAddressTypeOptions } from '../../../../components/data/converters/address_type';
+import contactTypes from '../../../../components/data/static/contactTypes';
 
 const EditContactInfo = ({
   expanded,
