+++ src/views/UserEdit/sections/EditExtendedInfo/CreateResetPasswordControl/Modals/CreatePasswordModal.js
