+++ src/views/UserEdit/sections/EditExtendedInfo/CreateResetPasswordControl/Modals/WithCopyModal/WithCopyModal.js
