+++ src/views/UserEdit/sections/EditExtendedInfo/CreateResetPasswordControl/Modals/WithCopyModal/index.js
